requirements.txt
py-cord==2.6.1
requests==2.31.0
python-dotenv==1.0.0
audioop-lts==0.2.2